pyproject.toml
[project]
name = "autoglm-gui"
version = "1.0.2"
description = "Web GUI for AutoGLM Phone Agent - AI-powered Android automation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "suyiiyii" }]
keywords = ["autoglm", "phone-agent", "android", "automation", "ai", "gui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.124.0",
    "httpx[socks]>=0.28.1",
    "loguru>=0.7.3",
    "openai>=2.9.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.0.0",
    "python-socketio>=5.11.0",
    "uvicorn[standard]>=0.38.0",
]

[project.urls]
Homepage = "https://github.com/suyiiyii/AutoGLM-GUI"
Repository = "https://github.com/suyiiyii/AutoGLM-GUI"

[project.scripts]
autoglm-gui = "AutoGLM_GUI.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["AutoGLM_GUI", "phone_agent"]
force-include = {"scrcpy-server-v3.3.3" = "scrcpy-server-v3.3.3"}

[tool.hatch.build.targets.sdist]
include = ["AutoGLM_GUI/**/*", "phone_agent/**/*", "scrcpy-server-v3.3.3"]

[tool.hatch.build]
artifacts = ["AutoGLM_GUI/static/**/*"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
    "ruff>=0.14.9",
]
